work

+++ backend/src/main/java/com/ll/standard/util/Ut.java
@@ -232,8 +232,8 @@ public class Ut {
         }
 
         @SneakyThrows
-        public static long getFileSize(String filePath) {
-            return Files.size(Path.of(filePath));
+        public static int getFileSize(String filePath) {
+            return (int) Files.size(Path.of(filePath));
         }
 
         @SneakyThrows
